Extract common.StreamDebugLog from api.Client
So we can use it from the migrationmaster worker.
